+++ eventsheet_addons/drawing_canvas/drawing_canvas_behavior.gd
@@ -1,5 +1,6 @@
 ## @ace_tags(drawing, visual)
 ## @ace_category("Drawing Canvas")
+## @ace_requires(CanvasSurface, DrawingPrefabResource)
 @icon("res://eventsheet_addons/drawing_canvas/icon.svg")
 class_name DrawingCanvas
 extends Node