+++ eventsheet_addons/drawing_prefab_stamp/drawing_prefab_stamp.gd
@@ -1,6 +1,7 @@
 @tool
 ## @ace_tags(drawing, visual)
 ## @ace_category("Drawing Canvas")
+## @ace_requires(DrawingPrefabResource)
 @icon("res://eventsheet_addons/drawing_prefab_stamp/icon.svg")
 class_name DrawingPrefabStamp
 extends Node2D